+++ src/components/Home/index.js
@@ -6,7 +6,6 @@ import Grid from "@material-ui/core/Grid";
 import Typography from "@material-ui/core/Typography";
 import Container from "@material-ui/core/Container";
 import Link from "@material-ui/core/Link";
-import CardMedia from "@material-ui/core/CardMedia";
 import Box from "@material-ui/core/Box";
 import NavigateNextIcon from "@material-ui/icons/NavigateNext";
 import Breadcrumbs from "@material-ui/core/Breadcrumbs";
@@ -20,12 +19,11 @@ import { FirebaseContext } from "../Firebase";
 import API from "../../api";
 import Statistics from "./statistics";
 import NewsPage from "./newsPage";
+import {transformCountryData, transformGlobalData} from "../../transform"
 
 import * as styles from "./styles";
 import * as ROUTES from "../../constants/routes";
 import * as IMAGES from "../../medias/images";
-
-const moment = require("moment");
 
 const GLOBAL = "Worldwide";
 const COUNTRIES = "Countries";
@@ -109,7 +107,7 @@ const Home = () => {
           <CountryTitle titleName={titleName} classes={classes} />
         </Route>
 
-        <Paper className={classes.grow}>
+        <Paper>
           <Tabs
             indicatorColor="primary"
             textColor="primary"
@@ -128,7 +126,7 @@ const Home = () => {
         {error && <Alert severity="error">{error.message}</Alert>}
 
         <Fade in={showStatistics} mountOnEnter unmountOnExit>
-          <div>
+          <div style={{ overflowX: "hidden"}}>
             <Statistics
               summary={summary}
               total={total}
@@ -139,7 +137,7 @@ const Home = () => {
         </Fade>
 
         <Fade in={!showStatistics} mountOnEnter unmountOnExit>
-          <div>
+          <div style={{ overflowX: "hidden"}}>
             <NewsPage loading={newsLoading} newsList={news}/>
           </div>
         </Fade>
@@ -180,96 +178,4 @@ const Footer = () => (
   </Box>
 );
 
-const transformCountryData = (data) => {
-  let totalDeaths = data.map((item) => item.Deaths);
-  let totalRecovered = data.map((item) => item.Recovered);
-  let totalConfirmed = data.map((item) => item.Confirmed);
-  let newDeaths = new Array(data.length);
-  let newRecovered = new Array(data.length);
-  let newConfirmed = new Array(data.length);
-  newDeaths[0] = newConfirmed[0] = newRecovered[0] = 0;
-
-  for (let i = 0; i < data.length - 1; i++) {
-    newDeaths[i + 1] = totalDeaths[i + 1] - totalDeaths[i];
-    newConfirmed[i + 1] = totalConfirmed[i + 1] - totalConfirmed[i];
-    newRecovered[i + 1] = totalRecovered[i + 1] - totalRecovered[i];
-  }
-
-  let labels = data.map((item) => moment(item.Date).format("MM-DD"));
-
-  return {
-    labels: labels,
-    totalDeaths: totalDeaths,
-    newDeaths: newDeaths,
-    totalRecoveries: totalRecovered,
-    newRecoveries: newRecovered,
-    totalConfirmed: totalConfirmed,
-    newConfirmed: newConfirmed,
-  };
-};
-
-const transformGlobalData = (data) => {
-  let computeDailyData = (currentValue, index, array) => {
-    if(index == 0){
-      return 0
-    }
-    return currentValue - array[index-1]
-  }
-
-  let labels = Object.keys(data["cases"])
-  let totalConfirmed = Object.values(data["cases"])
-  let totalRecovered = Object.values(data["recovered"])
-  let totalDeaths = Object.values(data["deaths"])
-  let newConfirmed = totalConfirmed.map(computeDailyData)
-  let newRecovered = totalRecovered.map(computeDailyData)
-  let newDeaths = totalDeaths.map(computeDailyData)
-
-  let processedData = {
-    labels: labels,
-    totalConfirmed: totalConfirmed,
-    newConfirmed: newConfirmed,
-    totalRecoveries: totalRecovered,
-    newRecoveries: newRecovered,
-    totalDeaths: totalDeaths,
-    newDeaths: newDeaths,
-  };
-  return processedData
-}
-
-/*
-const transformGlobalData = (data) => {
-  let start = moment("2020-04-12");
-  let sorting = function (a, b) {
-    return a[0] - b[0];
-  };
-
-  data.sort((a, b) => a.TotalConfirmed - b.TotalConfirmed);
-  let deaths = data
-    .map((item) => [item.TotalDeaths, item.NewDeaths])
-    .sort(sorting);
-  let totalDeaths = deaths.map((item) => item[0]);
-  let newDeaths = deaths.map((item) => item[1]);
-  let recovered = data
-    .map((item) => [item.TotalRecovered, item.NewRecovered])
-    .sort(sorting);
-  let totalRecovered = recovered.map((item) => item[0]);
-  let newRecovered = recovered.map((item) => item[1]);
-  let confirmed = data
-    .map((item) => [item.TotalConfirmed, item.NewConfirmed])
-    .sort(sorting);
-  let totalConfirmed = confirmed.map((item) => item[0]);
-  let newConfirmed = confirmed.map((item) => item[1]);
-  let labels = data.map((item) => start.add(1, "days").format("MM-DD"));
-
-  return {
-    labels: labels,
-    totalConfirmed: totalConfirmed,
-    newConfirmed: newConfirmed,
-    totalRecoveries: totalRecovered,
-    newRecoveries: newRecovered,
-    totalDeaths: totalDeaths,
-    newDeaths: newDeaths,
-  };
-};
-*/
 export default Home;
