+++ src/components/Navigation/styles.js
@@ -4,15 +4,12 @@ export const useStyles = makeStyles((theme) => ({
     grow: {
       flexGrow: 1,
     },
-    menuButton: {
-      marginRight: theme.spacing(2),
-    },
     title: {
       display: "none",
       [theme.breakpoints.up("sm")]: {
         display: "block",
+        marginRight: theme.spacing(2)
       },
-      marginRight: theme.spacing(2),
       textDecoration: 'none',
       "&:hover": {
         color: 'black',
@@ -36,16 +33,7 @@ export const useStyles = makeStyles((theme) => ({
     inputRoot: {
       color: "inherit",
     },
-    inputInput: {
-      padding: theme.spacing(1, 1, 1, 0),
-      // vertical padding + font size from searchIcon
-      paddingLeft: `calc(1em + ${theme.spacing(4)}px)`,
-      transition: theme.transitions.create("width"),
-      width: "100%",
-      [theme.breakpoints.up("md")]: {
-        width: "20ch",
-      },
-    },
+    
     sectionDesktop: {
       display: "none",
       [theme.breakpoints.up("md")]: {
